% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{AdjustPopulation}
\alias{AdjustPopulation}
\title{Adjust population vector to a new total}
\usage{
AdjustPopulation(orig.pop, new.total.pop)
}
\arguments{
\item{orig.pop}{original population vector}

\item{new.total.pop}{new total population size}
}
\value{
new population vector of class integer
}
\description{
Adjust population proportionally to a new total population size, rounding up
when necessary.
}
\keyword{internal}

