% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_amss_sim.R
\name{amss.sim}
\alias{amss.sim}
\title{Create AMSS simulation objects.}
\usage{
amss.sim(data = NULL, data.full = NULL, params = NULL)
}
\arguments{
\item{data}{observed data}

\item{data.full}{full data, as list of data.tables with rows corresponding
to each segment and columns corresponding to specific variables. each
data.table corresponds to a single time point.}

\item{params}{parameters used to generate the simulation}
}
\value{
object of class \code{amss.sim}, containing the observed data, the
  full dataset, and a prediction function.
}
\description{
Creates objects of class \code{amss.sim}, containing full information
about simulated data.
}

