% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.R
\name{ApplyTransitionMatrix}
\alias{ApplyTransitionMatrix}
\title{Apply transition matrix to vector of population counts by state.}
\usage{
ApplyTransitionMatrix(vector.counts, transition.matrices)
}
\arguments{
\item{vector.counts}{vector of counts in each state}

\item{transition.matrices}{transition matrix, or list of transition
matrices. When multiple transition matrices are supplied, the first
transition matrix whose dimensionality matches the number of states
in vector.counts is used.}
}
\value{
number of individuals in each state after migration
}
\description{
Applies transition matrix to vector of population counts by state.
Option to select transition matrix that matches length of vector when given
multiple transition matrices.
}
\keyword{internal}

