% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{CalculateROAS}
\alias{CalculateROAS}
\title{Calculate ROAS or mROAS.}
\usage{
CalculateROAS(object, new.budget = NULL,
  media.names = object$params$media.names, budget.periods = NULL,
  budget.proportion = rep(0, length(media.names)), t.start = 1,
  t.end = object$params$time.n, scaled.pop.size = .Machine$integer.max/100,
  min.reps = 10, max.coef.var = 0.01, max.margin.error = 0.01,
  max.time = 30, verbose = FALSE)
}
\arguments{
\item{object}{amss.sim object containing simulated data}

\item{new.budget}{table of new budgets for each budget period (row) and
media channel (column)}

\item{media.names}{if new.budget is NULL, adjust original budget of the
media named here.}

\item{budget.periods}{budget.periods over which to modify the budget.
Default \code{NULL} will lead to all budget periods being modified.}

\item{budget.proportion}{nonnegative numeric. When \code{new.budget} is
NULL, it is calculated by setting the budget of the media channels
specified in \code{media.names} to \code{budget.proportion}
proportion of the original budget during the budget periods specified
in \code{budget.periods}. The default proportion of 0 is used to
calculate the average ROAS over the entire spend in the channel.
Values such as 0.99 can be used to calculate the marginal ROAS.}

\item{t.start}{time point to start generating data according to the new
settings.}

\item{t.end}{last time point to generate data according to the new settings.
In scenarios with lag, this should extend past the last time point in
the modified budget periods in order to include lagged effects in the
calculation.}

\item{scaled.pop.size}{\code{CalculateROAS} scales up the population size to
reduce the variability of its estimates. This number should be chosen
to be as large as possible while avoiding integer overflow during
data simulation.}

\item{min.reps}{integer representing the initial number of datasets to
generate from each budget setting. A reasonable number of initial
datasets is needed to estimate the amount of variability accurately}

\item{max.coef.var}{numeric, the target coefficient of variation. The
function takes additional samples of the ROAS until it runs out of
time, attains the target coefficient of variation, or attains the
target margin of error.}

\item{max.margin.error}{numeric, the target margin of error. The function
takes additional samples of the ROAS until it runs out of time,
attains the target coefficient of variation, or attains the target
margin of error.}

\item{max.time}{numeric, the number of minutes at which to cut off the
function from taking additional samples beyond the initial sample
generated according to \code{min.reps}. The function takes additional
samples of the ROAS until it runs out of time, attains the target
coefficient of variation, or attains the target margin of error.}

\item{verbose}{boolean. If TRUE, output measures of the accuracy of the
reported ROAS, including the full sample of ROAS values.}
}
\value{
numeric value for ROAS, or, if \code{verbose = TRUE}, a list with
  the roas, the 95% margin of error, the coefficient of variation, and the
  sample ROAS values.
}
\description{
This functions takes the original budget settings and a
counterfactual budget setting. It reports the expected ratio between
the total difference in revenue over all time points and the total
difference in media spend over all time points.
}

