% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{CalculateSampleROAS}
\alias{CalculateSampleROAS}
\title{Calculate ROAS from a pair of datasets.}
\usage{
CalculateSampleROAS(dt1, dt2)
}
\arguments{
\item{dt1}{data.table of data generated from the first budget setting. The
data.table must include "revenue" and "total.spend" columns, and have
the different datasets indexed by column "rep.index."}

\item{dt2}{data.table of data generated from the second budget setting. The
data.table must include "revenue" and "total.spend" columns, and have
the different datasets indexed by column "rep.index." the number of
unique values of "rep.index" in \code{dt1} and \code{dt2} should
match.}
}
\value{
vector of ROAS estimates from each pair of sample datasets in dt1
  and dt2.
}
\description{
Calculate the ROAS estimated from a sample of datasets generated from two
budget settings
}
\keyword{internal}

