% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper.R
\name{Capitalize}
\alias{Capitalize}
\title{Capitalize the first letter of a string.}
\usage{
Capitalize(s)
}
\arguments{
\item{s}{the string to be capitalized.}
}
\value{
the string \code{s} with its first letter capitalized.
}
\description{
Capitalize the first letter of a string. This function is meant for single
strings only.
}
\keyword{internal}

