% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper.R
\name{CheckLength}
\alias{CheckLength}
\title{Check and adjust object length.}
\usage{
CheckLength(x, len, hard.stop = TRUE, warn.single = FALSE,
  par.name = character())
}
\arguments{
\item{x}{object to check/adjust length of}

\item{len}{desired length of object}

\item{hard.stop}{if TRUE, stop when length(x) is neither 1 nor len. else,
throw a warning and use repeat() to adjust length of x.}

\item{warn.single}{if TRUE, through a warning when length(x) is 1 and it is
being extended to a new length.}

\item{par.name}{Character, used in the warning or stop message to indicate
which parameter this warning is meant for.}
}
\value{
x updated to have length len, or throws exception
}
\description{
Function that checks length of x. It may throw a stop(), warning(), and/or
repeat the variable to create the desired length.
}
\keyword{internal}

