% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper.R
\name{CheckListNames}
\alias{CheckListNames}
\title{Check the names of a list.}
\usage{
CheckListNames(l, valid.names = colnames(kAllStates))
}
\arguments{
\item{l}{the list to be checked.}

\item{valid.names}{a character vector of valid nmaes.}
}
\value{
This function checks the value of \code{l} and will signal an error
  if the check fails. Else, it returns \code{TRUE}.
}
\description{
Check that a list has names, and that all these names are valid.
}
\keyword{internal}

