% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_sales.R
\name{CheckSalesActivity}
\alias{CheckSalesActivity}
\title{Warn users of possibility of consumers outside the 'purchase' state
purchasing.}
\usage{
CheckSalesActivity(x)
}
\arguments{
\item{x}{the parameters being checked}
}
\value{
\code{NULL}. If the parameter specification breaks enforcement of
  only consumers who have attained the 'purchase' state being able to make a
  purchase, the function signals a warning.
}
\description{
Checks parameters in the sales module to make sure that the probability of
consumers who are not in the 'purchase' activity state is 0.
}
\keyword{internal}

