% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_media.R
\name{DefaultSearchMediaModule}
\alias{DefaultSearchMediaModule}
\title{Model paid and/or organic search.}
\usage{
DefaultSearchMediaModule(data.dt, budget.index, budget,
  spend.cap.fn = function(time, budget, budget.indices) {     Inf },
  bid.fn = function(time, per.capita.budget, budget.indices) {     Inf },
  kwl.fn = function(time, per.capita.budget, budget.indices) {     1 },
  audience.membership = list(), query.rate = 1, cpc.min = 0,
  cpc.max = 1, ctr = list(), relative.effectiveness = c(0, 0, 1),
  transition.matrices = list())
}
\arguments{
\item{data.dt}{data.table with rows corresponding to population segments and
columns corresponding to specific variables}

\item{budget.index}{vector specifying budget period each time point belongs
to. For example, rep(1:4, each = 52) would correspond to 4 years of
yearly budget periods.}

\item{budget}{vector specifying the target spend for each budget period. For
example, given the example \code{budget.index} from above,
\code{budget = rep(1e6, 4)} would specify a budget of 1 million for
each year.}

\item{spend.cap.fn}{function mapping the current time, the budget, and the
budget period to a spend cap for the current week. By default this is
set to \code{Inf}, representing uncapped spend.}

\item{bid.fn}{function mapping the current time, the per-capita budget over
the population, and the budget period to a bid for the current week.
By default this is set to \code{Inf}, so that the advertiser wins all
auctions and will pay the maximum CPC.}

\item{kwl.fn}{function mapping the current time, the per-capita budget over
the population, and the budget period to the proportion of queries.
that match the keyword list. By default this is the maximum value of
1. To specify the proportion of matching queries by population
segment, have kwl.fn return a vector with entries for each segment.}

\item{audience.membership}{list of multipliers used to calculate probability
of audience membership. Each element of the list corresponds to a
specific dimension of population segmentation. Multipliers
corresponding to each dimension are multiplied to derive audience
membership probability for each segment. A named list with members
'activity', 'favorability', 'loyalty', and 'availability' is
expected. Each member is a numeric vector containing the multipliers
to use for each state in the dimension. For example, if member
"activity" is c(1, 0.5, 0.7), a multiplier of 0.7 should be used for
all segments with activity state "purchase." By default, any missing
members will have no effect.}

\item{query.rate}{nonnegative numeric, or vector. Each member of the
audience makes matching queries according to a Poisson process with
this rate. A vector rate specifies the query rate at each time. Note
that rate is the expected number of queries per person in the
audience. Defaults to 1. Vector repeats as necessary, so that
repeating patterns can be specified more simply.}

\item{cpc.min}{minimum CPC, defaults to 1. Must be nonnegative. vector
values are interpreted as the vector of minimum CPC's over time.}

\item{cpc.max}{maximum CPC. Must be at least as large as cpc.min. vector
values are interpreted as the vector of maximum CPC's over time.}

\item{ctr}{list of multipliers for each dimension with an effect on the
clickthrough rate (ctr). Values in each state are multiplied to
derive the ctr for each population segment. A named list with members
'activity', 'favorability', 'loyalty', and 'availability' is
expected. Each member is a numeric vector of the values for each
state in that dimension. By default, any missing members will have no
effect.}

\item{relative.effectiveness}{effectiveness, relative to the maximum
effectiveness specified by the transition matrices, by volume type:
organic only, paid impressions w/o paid click (click on organic
result included), and paid clicks. Default to maximum (1)
effectiveness for paid clicks, and no effect otherwise.}

\item{transition.matrices}{list of transition matrices for each dimension of
population segmentation that may be affected by marketing
interventions. A named list with members 'activity', 'favorability',
'loyalty', and 'availability' is expected. By default, any missing
members will have no effect.}
}
\value{
\code{invisible(NULL)}. \code{data.dt} updated by reference.
}
\description{
Simulate the behavior of a paid and/or organic search, including observable
variables (e.g., query volume, paid clicks, spend) and the effect on
consumer mindset.
}

