% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_media.R
\name{DefaultTraditionalMediaModule}
\alias{DefaultTraditionalMediaModule}
\title{Model the effect of a traditional media channel.}
\usage{
DefaultTraditionalMediaModule(data.dt, budget.index, budget,
  audience.membership = list(), flighting = rep(1, length(budget.index)),
  unit.cost = 1, effectiveness.function = NULL, hill.ec = 1,
  hill.slope = 1, transition.matrices = list())
}
\arguments{
\item{data.dt}{data.table with rows corresponding to population segments and
columns corresponding to specific variables}

\item{budget.index}{vector specifying budget period each time point belongs
to. For example, rep(1:4, each = 52) would correspond to 4 years of
yearly budget periods.}

\item{budget}{vector specifying the target spend for each budget period. For
example, given the example \code{budget.index} from above,
\code{budget = rep(1e6, 4)} would specify a budget of 1 million for
each year.}

\item{audience.membership}{list of multipliers used to calculate probability
of audience membership. Each element of the list corresponds to a
specific dimension of population segmentation. Multipliers
corresponding to each dimension are multiplied to derive audience
membership probability for each segment. A named list with members
'activity', 'favorability', 'loyalty', and 'availability' is
expected. Each member is a numeric vector containing the multipliers
to use for each state in the dimension. For example, if member
"activity" is c(1, 0.5, 0.7), a multiplier of 0.7 should be used for
all segments with activity state "purchase." By default, any missing
members will have no effect.}

\item{flighting}{specifies the relative amount to be spent on each time
point within a budget period. For example, in a budget period of two
weeks, \code{flighting = c(1,2)} specifies that twice 1/3 of the
budget should be spent in the first week, and 2/3 in the second.}

\item{unit.cost}{positive numeric specifying expected unit cost per
exposure.}

\item{effectiveness.function}{vectorized function mapping frequency to media
effect (relative to transition matrices specifying maximum effect).
The range of the function should be bounded between 0 and 1. Given
the default value of NULL, the module will used the Hill
transformation with parameters hill.ec and hill.slope.}

\item{hill.ec}{parameter controlling the scaling of frequency vs. effect.
This is the EC50 of the Hill transformation.}

\item{hill.slope}{parameter controlling the scaling of frequency vs. effect.
This is the maximum slope of the Hill transformation.}

\item{transition.matrices}{list of transition matrices for each dimension of
population segmentation that may be affected by marketing
interventions. A named list with members 'activity', 'favorability',
'loyalty', and 'availability' is expected. By default, any missing
members will have no effect.}
}
\value{
invisible(NULL). \code{data.dt} updated by reference.
}
\description{
Simulate the behavior of a traditional media channel, and generate
associated observable variables such as media volume and spend.
}

