% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_natural_migration.R
\name{Desatiate}
\alias{Desatiate}
\title{Implement desatiation.}
\usage{
Desatiate(data.dt, sat.decay, prop.activity = rep(1/length(kActivityStates),
  length(kActivityStates)))
}
\arguments{
\item{data.dt}{data.table to update}

\item{sat.decay}{rate for the geometric decay of satiation}

\item{prop.activity}{proportion of population assigned to each activity
state, given that they are responsive.}
}
\value{
\code{invisible(NULL)}. \code{data.dt} is updated by reference.
}
\description{
Implement desatiation.
}
\note{
Satiation has geometric decay with rate \code{sat.decay}. Thus,
during any time interval, \code{sat.decay} proportion of currently
satiated individuals desatiate. Individuals put into a responsive state
by desatiation are assigned activity states at rates defined by
prop.activity.
}
\keyword{internal}

