% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribution.R
\name{GenerateDataUnderNewBudget}
\alias{GenerateDataUnderNewBudget}
\title{Generate data under a new budget}
\usage{
GenerateDataUnderNewBudget(object, new.budget = GetBudget(object),
  response.metric = NULL, reps = 1, t.start = 1,
  t.end = object$params$time.n)
}
\arguments{
\item{object}{object of class \code{amss.sim} to do prediction on.}

\item{new.budget}{new budget levels for each media}

\item{response.metric}{string specifying what observable value to predict.
defaults to NULL, which results in entire dataset. Values such as
"log(brand.sales)" will make the prediction function return the
average log(brand.sales) over all reps, for ex.}

\item{reps}{number of replicates to generate}

\item{t.start}{time point at which to start generating new data}

\item{t.end}{time point at which to stop generating new data}
}
\value{
The list of all generated datasets or a vector of averaged values.
}
\description{
Simulate multiple datasets from a counterfactual simulation setting with
new budget settings.
}
\keyword{internal}

