% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_consts.R
\name{GetBrandStates}
\alias{GetBrandStates}
\title{Generate the list of brand states.}
\usage{
GetBrandStates()
}
\value{
\code{data.frame} with column \code{brand} containing all valid
  values for brand state.
}
\description{
Brand states are a combination of brand favorability, brand loyalty, and
brand availability. This functions produces a list all valid combinations
of these three dimensions of population segmentation.
}
\keyword{internal}

