% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_amss_sim.R
\name{GetBudget}
\alias{GetBudget}
\title{Get budget information from a simulation object.}
\usage{
GetBudget(object)
}
\arguments{
\item{object}{object of class amss.sim}
}
\value{
matrix of budgets by budget period (row) and media name (column)
}
\description{
Retrieves information on media budgets from an object of class
\code{amss.sim}.
}
\keyword{internal}

