% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_amss_sim.R
\name{GetBudgetIdx}
\alias{GetBudgetIdx}
\title{Get the budget period assigned to each time interval.}
\usage{
GetBudgetIdx(object)
}
\description{
Read the budget indices from an \code{amss.sim} object.
}
\details{
The budget indices specify which time intervals belong to the same budget
period.
}

