% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_consts.R
\name{GetCategoryStates}
\alias{GetCategoryStates}
\title{Generate the list of category states.}
\usage{
GetCategoryStates()
}
\value{
\code{data.frame} with columns \code{market}, \code{satiation}, and
  \code{activity}, listing all valid combinations of market, satiation,
  and activity states
}
\description{
Category states are a combination of market, satation, and activity state.
This functions produces a list all valid combinations of these three
dimensions of population segmentation.
}
\keyword{internal}

