% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{GetInterior}
\alias{GetInterior}
\title{Find an interior point in a bounding box.}
\usage{
GetInterior(lower.bound, upper.bound, sum.lower.bound, sum.upper.bound)
}
\arguments{
\item{lower.bound:}{vector of lower bounds for each coordinate.}

\item{upper.bound:}{vector of upper bounds for each coordinate.}

\item{sum.lower.bound:}{numeric, lower bound on the sum of all coordinates.}

\item{sum.upper.bound:}{numeric, upper bound on the sum of all coordinates.}
}
\value{
\code{GetInterior} returns a vector representing an interior point in
  the set.
}
\description{
The current optimization tool requires an interior point of the
set we are optimizing over. We assume that the constraints consist of a
bounding box (lower and upper bounds on each coordinate) and bounds on the
vector sum.
}
\details{
This function finds an interior point of the set bounded in each coordinate
by lower bounds \code{lower.bound} and upper bounds \code{upper.bound}, and
with sum bounded between \code{sum.lower.bound} and \code{sum.upper.bound}.
It firsts modifies the constraints to be finite, and then finds the point x
along the line passing from lower.bound to upper.bound such that
sum(x) = (sum.lower.bound + sum.upper.bound) / 2.
}
\keyword{internal}

