% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_amss_sim.R
\name{GetPopulation}
\alias{GetPopulation}
\title{Get population size from a simulation object.}
\usage{
GetPopulation(object)
}
\arguments{
\item{object}{object of class amss.sim}
}
\value{
integer population size
}
\description{
Retrieves population size from an object of class \code{amss.sim}.
}
\keyword{internal}

