% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper.R
\name{HillTrans}
\alias{HillTrans}
\title{Define the Hill transformation function.}
\usage{
HillTrans(x, ec, slope, beta = 1)
}
\arguments{
\item{x}{original input.}

\item{ec}{effective concentration parameter.}

\item{slope}{slope parameter}

\item{beta}{vertical scale parameter, defaults to 1}
}
\value{
tranformed value = beta / (1 + (x / ec) ^ (-slope))
}
\description{
The Hill function is one option for parameterizing a flexible set of S-shaped
curves.
}
\keyword{internal}

