% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_natural_migration.R
\name{InitPop}
\alias{InitPop}
\title{Initialize population segmentation.}
\usage{
InitPop(data.dt, pop.total, market.rate = 1,
  prop.activity = rep(1/length(kActivityStates), length(kActivityStates)),
  prop.favorability = rep(1/length(kFavorabilityStates),
  length(kFavorabilityStates)), prop.loyalty = rep(1/length(kLoyaltyStates),
  length(kLoyaltyStates)),
  prop.availability = rep(1/length(kAvailabilityStates),
  length(kAvailabilityStates)))
}
\arguments{
\item{data.dt}{data.table containing all state-related data}

\item{pop.total}{total population}

\item{market.rate}{target proportion of consumers in 'in-market' market
state.}

\item{prop.activity}{vector of nonnegative values summing to 1, representing
the proportion of the population to be assigned to each activity
state, given they are "responsive," i.e., "in.market" and
"unsatiated."}

\item{prop.favorability}{vector of nonnegative values summing to 1,
representing the proportion of the population to be assigned to each
favorability state, given they are not "loyal."}

\item{prop.loyalty}{vector of nonnegative values summing to 1, representing
the proportion of the population to be assigned to each loyalty
state.}

\item{prop.availability}{vector of nonnegative values summing to 1,
representing the proportion of the population to be assigned to each
availability state.}
}
\value{
\code{invisible(NULL)}. \code{data.dt} is updated by reference.
}
\description{
Initialize population segmentation.
}
\keyword{internal}

