% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{InitStateData}
\alias{InitStateData}
\title{Initialize data.}
\usage{
InitStateData(time.index = 0)
}
\arguments{
\item{time.index}{time index to intialize the data table to}
}
\value{
data.table with one column for each state variable (market,
  satiation, activity, brand) and one row per each valid combination of
  states for the above state variables. Also initializes the population
  count at timepoint 1 as 0.
}
\description{
Initialize the updating \code{data.table} containing current data on each
population segment.
}
\keyword{internal}

