% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_consts.R
\docType{data}
\name{kActivityStates}
\alias{kActivityStates}
\title{Constant defining the activity states.}
\format{An object of class \code{character} of length 3.}
\usage{
kActivityStates
}
\description{
\code{kActivityStates} is a character vector of all valid activity states a
consumer may take. Activity state tracks what category-related activities
the consumer is engaged in.
}
\keyword{datasets}

