% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_consts.R
\docType{data}
\name{kAllStates}
\alias{kAllStates}
\title{Constant defining all consumer states.}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 198 rows and 6 columns.}
\usage{
kAllStates
}
\description{
A \code{data.table} of all valid consumer states. It is the cross product of
category and brand states. Every consumer is assigned to one of these 198
states.
}
\keyword{datasets}

