% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_consts.R
\docType{data}
\name{kAvailabilityStates}
\alias{kAvailabilityStates}
\title{Constant defining the brand availability states.}
\format{An object of class \code{character} of length 3.}
\usage{
kAvailabilityStates
}
\description{
\code{kAvailabilityStates} is a character vector of all valid brand
availability states a consumer may take. Brand availability may refer to
the physical availability of the advertiser's product to a particular
consumer, or to the mental availability, i.e., the convenience.
}
\keyword{datasets}

