% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_consts.R
\docType{data}
\name{kBrandStates}
\alias{kBrandStates}
\title{Constant defining the brand states.}
\format{An object of class \code{data.frame} with 33 rows and 3 columns.}
\usage{
kBrandStates
}
\description{
A \code{data.table} of all valid combinations of brand favorability state,
brand loyalty state, and brand availability state, given that only consumers
with a \code{favorable} opinion of the brand can be \code{loyal}. Brand
state summarizes the consumer's relationship with the advertiser's brand.
}
\keyword{datasets}

