% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_consts.R
\docType{data}
\name{kCategoryStates}
\alias{kCategoryStates}
\title{Constant defining the category states.}
\format{An object of class \code{data.frame} with 6 rows and 3 columns.}
\usage{
kCategoryStates
}
\description{
A \code{data.frame} of all valid combinations of market state, satiation
state, and activity state, given that only consumers who are both
\code{in.market} and \code{unsatiated} can have activity states other than
\code{inactive}. Category state summarizes the consumer's relationship with
the category (as opposed to brand-specific relationships).
}
\keyword{datasets}

