% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_consts.R
\docType{data}
\name{kFavorabilityStates}
\alias{kFavorabilityStates}
\title{Constant defining the brand favorability states.}
\format{An object of class \code{character} of length 5.}
\usage{
kFavorabilityStates
}
\description{
\code{kFavorabilityStates} is a character vector of all valid brand
favorability states a consumer may take. Brand awareness is also tracked
through this dimension; consumers are either "unaware" of the advertiser's
brand, or have a brand favorability ranging from "negative" to "favorable."
}
\keyword{datasets}

