% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_consts.R
\docType{data}
\name{kLoyaltyStates}
\alias{kLoyaltyStates}
\title{Constant defining the brand loyalty states.}
\format{An object of class \code{character} of length 3.}
\usage{
kLoyaltyStates
}
\description{
\code{kLoyaltyStates} is a character vector of all valid brand loyalty states
a consumer may take.
}
\keyword{datasets}

