% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_consts.R
\docType{data}
\name{kMarketStates}
\alias{kMarketStates}
\title{Constant defining the market states.}
\format{An object of class \code{character} of length 2.}
\usage{
kMarketStates
}
\description{
\code{kMarketStates} is a character vector of all valid market states a
consumer may take. Consumers who have no interest in the category should be
considered "out-of-market." Consumers' market states may vary over time
due to seasonal changes in demand, but generally should not be changed by
marketing interventions.
}
\keyword{datasets}

