% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_consts.R
\docType{data}
\name{kSatiationStates}
\alias{kSatiationStates}
\title{Constant defining the satiation states.}
\format{An object of class \code{character} of length 2.}
\usage{
kSatiationStates
}
\description{
\code{kSatiationStates} is a character vector of all valid satiation states a
consumer may take. Satiation tracks whether a consumer's demand for the
product category is temporarily satisfied by a recent purchase.
}
\keyword{datasets}

