% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.R
\name{MigrateMarginal}
\alias{MigrateMarginal}
\title{Simulate migration in a single dimension of population segmentation.}
\usage{
MigrateMarginal(data.dt, migrating.pop.size, migration.dim, transition.matrix)
}
\arguments{
\item{data.dt}{data.table with rows representing population segments and
columns representing specific variables.}

\item{migrating.pop.size}{migrating population size}

\item{migration.dim}{name dimension of migration, must be a column of
kAllStates.}

\item{transition.matrix}{transition matrix specifying probabilities of
migration between states.}
}
\value{
\code{invisible(NULL)}. \code{data.dt} is updated by reference.
}
\description{
Simulate migration of indivduals between different states in single
dimension of population segmentation, such as "brand favorability."
}
\keyword{internal}

