% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.R
\name{MigrateMultiple}
\alias{MigrateMultiple}
\title{Simulate migration in multiple dimensions of population segmentation.}
\usage{
MigrateMultiple(data.dt, migrating.pop.size, migration.dims = character(),
  transition.matrices = list())
}
\arguments{
\item{data.dt}{data.table with rows representing population segments and
columns representing specific variables.}

\item{migrating.pop.size}{migrating population size}

\item{migration.dims}{a character vector of dimensions of migration, by
name.}

\item{transition.matrices}{a list of transition matrices for each dimension.}
}
\value{
\code{invisible(NULL)}. \code{data.dt} is updated by reference.
}
\description{
Perform successive migrations of consumers between population segments, with
each migration focusing on changes in a particular dimension of population
segmentation.
}
\keyword{internal}

