% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper.R
\name{MultiplyBySegment}
\alias{MultiplyBySegment}
\title{Calculate the product of population segmentation dimension-specific factors.}
\usage{
MultiplyBySegment(multiplicand.list = list(), starting.multiplicand = 1)
}
\arguments{
\item{multiplicand.list}{list of named numeric vectors. Each vector
corresponds to the named dimension of population segmentation. It
specifies the value assigned to every state possible in that
dimension. Dimensions other than 'activity', 'favorability',
'loyalty', and 'availailability' are ignored. Dimensions missing from
the list do not affect the product.}

\item{starting.multiplicand}{numeric constant, additional factor to multiply
every product by. Default 1.}
}
\value{
numeric vector. For each population segment, the vector holds the
  product of multiplying all factors corresponding to that segment.
}
\description{
Multiplies factors corresponding to dimensions of segmentation.
}
\keyword{internal}

