% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper.R
\name{ParseT}
\alias{ParseT}
\title{Parse text string.}
\usage{
ParseT(text)
}
\arguments{
\item{text}{string to be sent as \code{text} argument to \code{parse}.}
}
\value{
the text as an expression, result of calling
  \code{parse(text = text)}.
}
\description{
Shorthand for parse(text = ...), frequently used with data.table This
saves us from having to redo row selection in the env argument of
EvalText().
}
\note{
Including outer \code{eval()} call in the shorthand caused errors,
and needs to be done separately.
}
\examples{
dt <- data.table::as.data.table(mtcars)
colname <- "mpg"
dt[1:5, eval(amss:::ParseT(colname))]
dt[, (amss:::PasteD(colname, "plus1")) := eval(amss:::ParseT(colname)) + 1]

}
\keyword{internal}

