% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper.R
\name{PasteD}
\alias{PasteD}
\title{Paste with "." separator.}
\usage{
PasteD(...)
}
\arguments{
\item{...}{vector of strings to concatenate.}
}
\value{
concatenated string.
}
\description{
Function to simplify pasting with . as separator.
}
\keyword{internal}

