% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation_consts.R
\name{population segmentation}
\alias{kActivityStates}
\alias{kAllStates}
\alias{kAvailabilityStates}
\alias{kBrandStates}
\alias{kCategoryStates}
\alias{kFavorabilityStates}
\alias{kLoyaltyStates}
\alias{kMarketStates}
\alias{kSatiationStates}
\alias{population segmentation}
\title{Population segmentation constants}
\format{An object of class \code{character} (all possible states in a single
  dimension) or \code{data.table} (each row specifying a valid combination of
  states in different dimensions).}
\description{
AMSS segments the population segments the population into groups based
on each consumer's current mindset with regards to the category and the
advertiser's brand. Aggregate changes in the population are tracked through
the size of, i.e., the number of individuals belonging to, each population
segment.
}
\details{
The consumer minset is defined along six dimensions. The first three specify
the consumer's relationship with the category:
\describe{
  \item{Market state}{describes whether the consumer should be
    considered part of the market for this category. Consumers with no
    interest are \code{out.of.market}; the rest are \code{in.market}. A
    consumer's market state may vary over time due to seasonal changes in
    consumer demand, but generally should not be affected by marketing
    interventions.}
  \item{Satiation state}{tracks whether a consumer's demand for the
    product category is temporarily satisfied by a recent purchase.}
  \item{Activity state}{tracks the consumer's progress along the path to
    purchase. Consumers may be in the \code{inactive}, \code{exploratory}, or
    \code{purchase} state. Consumers in different activity states will have
    different behaviors. For example, by default consumers outside the
    \code{purchase} state will never make a purchase. Activity state also
    affects media consumption; for example, individuals who are not
    \code{inactive} are generally more likely to make generic or branded
    search queries.}
}
The last three dimensions describe the consumer's relationships with the
advertiser's brand.
\describe{
  \item{Brand favorability state}{specifies a consumer's awareness of and
    opinion of the advertiser's brand. Consumers are either \code{unaware},
    or are aware and have an opinion of the brand ranging from
    \code{negative} to \code{favorable}, with intermediate favorabilitiy
    levels \code{neutral} and \code{somewhat favorable}.}
  \item{Brand loyalty state}{specifies a consumer's loyalty status. A
    consumer may be a \code{switcher}, in which case he or she has no brand
    loyalty. Otherwise the consumer is either \code{loyal}, i.e., loyal to
    the advertiser's brand, or \code{competitor.loyal}.}
  \item{Brand availability state}{refers to whether the advertiser's product
    is easily available to a particular consumer. For example, if the
    advertiser's distribution efforts only cover seventy percent of the
    population, then the thirty percent of the population not covered would
    be in the \code{low} brand availability state. The other options are
    \code{average} and \code{high} brand availability. Availability can
    refer to physical availability, i.e. the presence of the advertiser's
    product on store shelves. It could also refer to the mental availability
    (convenience) of the advertiser's brand. Thus brand availability can
    be affected by, say search ads that make the advertiser's brand the most
    prominent on the search results page, or by having the advertiser's
    product at eye-level in a store shelf.}
}

The constants \code{kMarketStates}, \code{kSatiationStates},
\code{kActivityStates}, \code{kFavorabilityStates}, \code{kLoyaltyStates},
and \code{kAvailabilityStates} list the the possible states a consumer may
take in each dimension as character vectors.

A consumer's mindset is summarized by the combination of states they take
in each dimension. There are certain restrictions on which combinations of
consumer states are possible. For example, only consumers who are both
\code{in.market} and \code{unsatiated} can leave the \code{inactive} activity
state. The \code{data.frame} \code{kCategoryStates} describes all valid
combinations of market state, satiation state, and activity state, and thus
lists all possible consumer mindsets with respect to the category in general.
The \code{data.frame} \code{kBrandStates} describes all valid combinations of
brand favorability, loyalty, and availability, given that only consumers
with a \code{favorable} opinion of the brand can be \code{loyal}. Thus,
\code{kBrandStates} lists all possible consumer mindsets with regards to the
advertiser's brand.

A \code{data.table} of all valid consumer states is provided as
\code{kAllStates}. It is the cross product of all category and brand states.
Every consumer is assigned to one of these 198 states.
}

