% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_number_generator.R
\name{RBinom}
\alias{RBinom}
\title{Simulate random number from the binomial distribution.}
\usage{
RBinom(n, size, prob)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length is
taken to be the number required.}

\item{size}{number of trials (zero or more). When \code{size} is a vector,
the number of trials for each observation. Conflicts between the number of
observations \code{n} and the length of the size vector are resolved by
truncating or repeating it to length \code{n}. This matches the behavior of
the original \code{rbinom()}.}

\item{prob}{probability of success on each trial. When \code{prob} is a
vector, each entry refers to the probability of success for trials
associated with the corresponding observation. Conflicts between the
number of observations \code{n} and the length of the probability vector
are resolved by truncating or repeating it to length \code{n}. This matches
the behavior of the original \code{rbinom()}.}
}
\value{
Random numbers from the binomial distribution, or its normal
  approximation. The function will return a numeric value, rather than an
  integer.
}
\description{
Generate random integers from the binomial distribution when possible within
integer overflow constraints. Otherwise, approximate with the normal
distribution.
}
\keyword{internal}

