% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_helper.R
\name{ReadRepeatingVector}
\alias{ReadRepeatingVector}
\title{Read entry from repeated vector.}
\usage{
ReadRepeatingVector(v, idx)
}
\arguments{
\item{v}{the vector being read}

\item{idx}{the index to read the value from.}
}
\value{
the \code{idx}-th entry of \code{v}, under the assumption that
  \code{v} repeats as necessary to reach length \code{idx}.
}
\description{
Read an entry of a vector, under the assumption that it repeats to the
necessary length.
}
\keyword{internal}

