% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize.R
\name{ReduceDimension}
\alias{ReduceDimension}
\title{Reduce the dimensionality of the optimization}
\usage{
ReduceDimension(n.dim, lower.bound, upper.bound, sum.lower.bound,
  sum.upper.bound)
}
\arguments{
\item{n.dim}{integer, original dimensionality of the optimization.}

\item{lower.bound}{numeric vector, the constraint specifying the lower bound
in each dimension.}

\item{upper.bound}{numeric vector, the constraint specifying the upper bound
in each dimension.}

\item{sum.lower.bound}{numeric, the lower bound on the vector sum.}

\item{sum.upper.bound}{numeric, the upper bound on the vector sum.}
}
\value{
\code{ReduceDimension} returns \code{NULL} if the constrained set
  is empty. Else, it returns a code{list} with elements
  \describe{
    \item{lower.bound}{the new lower bound.}
    \item{upper.bound}{the new upper bound.}
    \item{sum.lower.bound}{the new lower bound on the vector sum.}
    \item{sum.upper.bound}{the new upper bound on the vector sum.}
    \item{decoder}{function mapping values in the new search space to values
      in the original search space.}
    \item{encoder}{function mapping values in the original search space to
      values in the new search space.}
  }
}
\description{
Reduce the dimensionality of the optimization and find a function that
maps the new search space of the optimization to the original search space.
}
\details{
The search space is assumed to be the intersection of a bounding box and an
upper/lower bound on the vector sum. If possible, the function finds an
equivalent lower-dimensional search space of the same type, and provides a
function mapping the new search space to the old search space.
}
\keyword{internal}

