% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_number_generator.R
\name{RHyper}
\alias{RHyper}
\title{Simulate random number from the hypergeometric distribution.}
\usage{
RHyper(nn, m, n, k)
}
\arguments{
\item{nn}{number of observations. If \code{length(nn) > 1}, the length is
taken to be the number required.}

\item{m}{the number of white balls in the urn.}

\item{n}{the number of black balls in the urn.}

\item{k}{the number of balls drawn from the urn.}
}
\value{
Random numbers from the binomial distribution, or its normal
  approximation. The function will a numeric value, rather than an integer.
}
\description{
Generate random integers from the hypergeometric distribution when possible
within integer overflow constraints. Otherwise, approximate with the normal
distribution.
}
\details{
The hypergeometric distribution is used for sampling the number of white
balls drawn when a fixed number of balls is drawn without replacement from
an urn which contains both black and white balls.
}
\keyword{internal}

