% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_number_generator.R
\name{RMultinom}
\alias{RMultinom}
\title{Simulate random number from the multinomial distribution.}
\usage{
RMultinom(n, size, prob)
}
\arguments{
\item{n}{the number of random vectors to draw}

\item{size}{integer, say \eqn{N}, specifying the total number of objects
that are put into \eqn{K} boxes in the typical multinomial experiment.}

\item{prob}{numeric non-negative vector of length \eqn{K}, specifying the
probability for the \eqn{K} classes; is internally normalized to sum 1.
Infinite and missing values are not allowed.}
}
\value{
Matrix of random numbers from the multinomial distribution, or its
  normal approximation. The function will return a numeric value, rather than
  an integer.
}
\description{
Generate random integers from the binomial distribution when possible within
integer overflow constraints. Otherwise, approximate with the normal
distribution.
}
\keyword{internal}

