% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_number_generator.R
\name{RPois}
\alias{RPois}
\title{Simulate random number from the Poisson distribution.}
\usage{
RPois(n, lambda)
}
\arguments{
\item{n}{number of random values to return. If \code{length(n) > 1}, the
length is taken to be the number required.}

\item{lambda}{vector of (non-negative) means.}
}
\value{
RPois returns random numbers from the Poisson distribution, or its
  normal approximation. The function will return a numeric value, rather than
  an integer.
}
\description{
Generate random integers from the Poisson distribution if possible, given
integer overflow constraints. Otherwise, approximate the Poisson with the
normal distribution.
}
\keyword{internal}

