% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_time_series.R
\name{SimulateAR1}
\alias{SimulateAR1}
\title{Simulate AR1 time series}
\usage{
SimulateAR1(n, stable.mu = 0, stable.sd = 1, autocor = 0)
}
\arguments{
\item{n}{integer number of time points}

\item{stable.mu}{means of the stable distribution for each variable}

\item{stable.sd}{standard deviations of the stable distributions}

\item{autocor}{autocorrelations for each time series}
}
\value{
vector realization of specified AR1 times series
}
\description{
Function that outputs simulated AR1 time series with specified means,
variances, and autocorrelations
}
\seealso{
Other Simulate.time.series: \code{\link{SimulateCorrelated}},
  \code{\link{SimulateDummy}},
  \code{\link{SimulateSinusoidal}}
}

