% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_time_series.R
\name{SimulateCorrelated}
\alias{SimulateCorrelated}
\title{Simulate correlated vectors.}
\usage{
SimulateCorrelated(v, cor.vx = 1, mu.x = 0, sigma.x = 1)
}
\arguments{
\item{v}{vector to which the new data will be correlated.}

\item{cor.vx}{numeric specifying correlation between v and the new vector x.}

\item{mu.x}{numeric, mean of the new vector x.}

\item{sigma.x}{numeric, standard deviation of the new vector x.}
}
\value{
a vector x with the specified mean, standard deviation, and
  correlation.
}
\description{
Simulates a new vector x with a specified mean, standard deviation, and
correlation with some other vector \code{v} by adding white noise and
scaling.
}
\seealso{
Other Simulate.time.series: \code{\link{SimulateAR1}},
  \code{\link{SimulateDummy}},
  \code{\link{SimulateSinusoidal}}
}

