% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{SimulateData}
\alias{SimulateData}
\title{Simulate data using the AMSS framework.}
\usage{
SimulateData(starting.dts = list(), time.n, nat.mig.module, nat.mig.params,
  media.names, media.modules, media.params, sales.module, sales.params, ping)
}
\arguments{
\item{starting.dts}{previously generated data that we can start with. data
will only be generated for later timepoints}

\item{time.n}{total number of timepoints, including the timepoints already
existing in starting.dts.}

\item{nat.mig.module}{function used to simulate effects of natural migration
on population segmentation.}

\item{nat.mig.params}{list of any parameter values to pass to
\code{nat.mig.module}}

\item{media.names}{character vector of unique names of all media modules.
ex: c("tv", "search")}

\item{media.modules}{list of functions that simulate the behavior of each
marketing intervention.}

\item{media.params}{list of parameter value lists for each media module.}

\item{sales.module}{function that generate sales variables.}

\item{sales.params}{list of any parameter values to pass to the sales module}

\item{ping}{the spacing between time points at which to print a message
updating the user on simulation progress.}
}
\value{
a list of data sets, one per timepoint. Each is a data.table with
  rows corresponding to population segments and columns corresponding
  to specific variables.
}
\description{
Data generating function for AMSS. Responsible for sequentially simulating
the value of each variable at each time point by population segment, given
the simulation settings.
}
\keyword{internal}

