% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_time_series.R
\name{SimulateDummy}
\alias{SimulateDummy}
\title{Simulate dummy (0-1) variables.}
\usage{
SimulateDummy(n, pos.idx = NULL, period = n, amplitude = 1)
}
\arguments{
\item{n}{integer number of time points}

\item{pos.idx}{vector of indices where simulated vector should take positive
values (as opposed to zero)}

\item{period}{integer controlling periodicity.}

\item{amplitude}{numeric value > 0 specifying the value of all postive
entries in the return vector.}
}
\value{
specified vector
}
\description{
Create dummy (0-1) variables that repeat a requested pattern of 0's and 1's,
with the option to scale.
}
\seealso{
Other Simulate.time.series: \code{\link{SimulateAR1}},
  \code{\link{SimulateCorrelated}},
  \code{\link{SimulateSinusoidal}}
}

