% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_number_generator.R
\name{SimulateNotEmptyUrns}
\alias{SimulateNotEmptyUrns}
\title{Simulate the number of urns.}
\usage{
SimulateNotEmptyUrns(m, n, exact.n = 20)
}
\arguments{
\item{m}{the number of balls, single integer or vector of integers.}

\item{n}{the number of urns, single integer or vector of integers.}

\item{exact.n}{single integer, the maximum number of urns for which to use
exact calculations instead of a normal approximation.}
}
\value{
the simulated number of non-empty urns
}
\description{
Simulate the number of non-empty urns when m balls placed into n urns.
}
\keyword{internal}

