% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_time_series.R
\name{SimulateSinusoidal}
\alias{SimulateSinusoidal}
\title{Generate sinusoidal time series.}
\usage{
SimulateSinusoidal(n, period, max.loc = 1, vert.translation = 0,
  amplitude = 1, scale.x = FALSE)
}
\arguments{
\item{n}{the length of the simulated vector.}

\item{period}{the length of one full sinusoidal period.}

\item{max.loc}{the index of the maximum of the sinusoidal curve.}

\item{vert.translation}{a numeric for the vertical displacement of the
sinusoidal curve from 0.}

\item{amplitude}{numeric for the amplitude of the sinusoidal curve. Must be
nonnegative.}

\item{scale.x}{boolean. If TRUE, scale the sinusoidal curve to have mean 0
and standard deviation 1 before returning.}
}
\value{
specified sinusoidal curve as a vector
}
\description{
Function that outputs specified sinusoidal waves.
}
\seealso{
Other Simulate.time.series: \code{\link{SimulateAR1}},
  \code{\link{SimulateCorrelated}},
  \code{\link{SimulateDummy}}
}

