% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{SurfaceData}
\alias{SurfaceData}
\title{Surface observable data.}
\usage{
SurfaceData(full.data, names.const, names.sum)
}
\arguments{
\item{full.data}{\code{data.table} of full data. each row contains
information for a particular timepoint + hidden state}

\item{names.const}{character names of variables with constant values over
the state, to be aggregated by taking the first entry.}

\item{names.sum}{character names of variables with numeric values, to be
aggregated by \code{sum()}.}
}
\value{
observable data, aggregated over the hidden states as specified.
  Variables not included in names.const or names.sum are assumed to be
  hidden variables and not surfaced.
}
\description{
Extract observable data from the full dataset.
}
\keyword{internal}

