% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_natural_migration.R
\name{UpdateMarket}
\alias{UpdateMarket}
\title{Updates in/out of market status by moving mimimal population in/out of the
market necessary to match on requested proportion of "in.market"
individuals.}
\usage{
UpdateMarket(data.dt, target.rate,
  prop.activity = rep(1/length(kActivityStates), length(kActivityStates)))
}
\arguments{
\item{data.dt}{data.table containing all state-related data}

\item{target.rate}{target proportion of consumers 'in-market'}

\item{prop.activity}{single value between 0 and 1, representing proportion
of population to be assigned to each activity state, given they are
"responsive," i.e., "in.market" and "unsatiated."}
}
\value{
\code{invisible(NULL)}. \code{data.dt} is updated by reference.
}
\description{
Updates in/out of market status by moving mimimal population in/out of the
market necessary to match on requested proportion of "in.market"
individuals.
}
\keyword{internal}

