% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_natural_migration.R
\name{UpdateMarketingResponsiveStates}
\alias{UpdateMarketingResponsiveStates}
\title{Update segmentation in marketing-responsive states (activity,
favorability, loyalty, and availability) according to specified
transition matrices.}
\usage{
UpdateMarketingResponsiveStates(data.dt, transition.matrices = list())
}
\arguments{
\item{data.dt}{data.table with column 'pop' for population segment size.}

\item{transition.matrices}{list of transition matrices for each dimension of
population segmentation that may be affected by marketing
interventions. A named list with members 'activity', 'favorability',
'loyalty', and 'availability' is expected. By default, any missing
members will have no effect.}
}
\value{
\code{invisible(NULL)}. \code{data.dt} is updated by reference.
}
\description{
Update segmentation in marketing-responsive states (activity,
favorability, loyalty, and availability) according to specified
transition matrices.
}
\keyword{internal}

