% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{SimulateAMSS}
\alias{SimulateAMSS}
\title{Generate simulation objects under the AMSS framework.}
\usage{
SimulateAMSS(time.n, geo.index = 0, nat.mig.module = DefaultNatMigModule,
  nat.mig.params = list(), media.names = character(),
  media.modules = rep(list(DefaultTraditionalMediaModule),
  length(media.names)), media.params = rep(list(list()), length(media.names)),
  sales.module = DefaultSalesModule, sales.params = list(), ping = max(10,
  floor(time.n/10)), names.agg.const = NULL, names.agg.sum = NULL)
}
\arguments{
\item{time.n}{number of timepoints.}

\item{geo.index}{single value, value of geo index for entire dataset.}

\item{nat.mig.module}{specifications of class \code{seq.specs}, to be used
to create the non-actionable drivers module and then generate its
variables}

\item{nat.mig.params}{any parameter values to pass to \code{nat.mig.module}}

\item{media.names}{character vector of unique names of all media modules.
ex: c("tv", "search")}

\item{media.modules}{list of functions that simulate the behavior of each
marketing intervention.}

\item{media.params}{list of parameter value lists for each media module.}

\item{sales.module}{function that models the sales in the category.}

\item{sales.params}{list of any parameter values to pass to the sales module}

\item{ping}{the spacing between time points at which to print a message
updating the user on simulation progress.}

\item{names.agg.const}{character vector of names of variables to surface in
the observed data, aggregated using the first entry, since they are
constant over the hidden states. By default, if not specified,
\code{SurfaceData()} will pick up variables with names containing
"price" and/or "budget", and the "pop.total" variable.}

\item{names.agg.sum}{character vector of names of variables to surface in
the observed data, aggregated using the function \code{sum()}. By
default, \code{SurfaceData()} will pick up variables with names
matching "revenue", "profit", "sales", "volume", and/or "spend".}
}
\value{
an object of class \code{amss}, containing
  \describe{
    \item{data:}{observed data.}
    \item{data.full:}{the full data, as a list of data.tables.}
    \item{params:}{the simulation settings used to generate the data.}
    }
}
\description{
Produces an \code{amss.sim} object that contains the simulated data and can
be used to derive ground truth about the scenario.
}

